Fix 404 page for webpack
