+++ app/404/index.html
@@ -30,7 +30,7 @@
   <script src='https://api.mapbox.com/mapbox.js/v2.4.0/mapbox.js'></script>
   <link href='https://api.mapbox.com/mapbox.js/v2.4.0/mapbox.css' rel='stylesheet' />
 
-  <script src="/404/404.js"></script>
+  <script src="/404/index.js"></script>
 </body>
 
 </html>
